BadWords (#7)
* Updated >dad with properly cropped image

* Added custom commands, moved some custom commands to bot.js and added a helper commands command for displaying custom commands.

* Fixed roll so it goes 1-100

* moved dad back to custom commands

* Fixed remindme

* Added ask command and joined the modules for ask, flipacoin and roll into 'FUN'

* Added commands from MEE6 and moved help from custom to main bot file

* Added the ability to add, remove and list custom commands with the 'commands' command

* Updated imported commands with active links

* Properly fixed reminders

* Commented unnecessary code in remindme

* Commented unnecessary code in remindme

* Added precautionary return statements

* Actually (most likely) fixed reminders

* Fixed cooldown for adding commands

* Removed embeds from custom commands

* Added confirmation messages for adding and removing commands.

* Added custom command for steam being down

* Rounded up ping in status command

* Fixed message for removing custom commands(ty Gab)

* Renamed variable

* Removed unnecessary variable

* Reworked help

* Reworked help so the embed is an object

* Fixed commands list

* Reworked badWords

* Removed commented checkReminders

* Minor fixes

* Minor fixes

* Ensuring the fixes are same as on master

* Small fixes

* Separated modules for ask, roll and flipacoin

* Changed the way badWords are stored

* Resolving conflicts

* Removed guild id from BadWords

+++ bot.js
@@ -540,7 +540,12 @@ function isNull(value, def){
 function checkMessageContent(message){
   if(message.member.roles.some(role => ["Moderators"].includes(role.name))) return;
   var wholeMessage = (message.content).split(" ");
-  var badWordList = badWordsFile.get(`${message.guild.id}.badWords`);
+  var badWordList = badWordsFile.get(`badWords`);
+  if (badWordList == undefined) {
+    badWordsFile.set(`badWords`, []);
+    badWordsFile.save();
+    return;
+  }
   if(badWordList.length > 0) {
     if(badWordList.some(word => wholeMessage.includes(word))){
       message.delete()
@@ -796,7 +801,7 @@ client.on('message', async message => {
   if(message.content.indexOf(config.prefix) !== 0) return; //If the message content doesn't start with our prefix, return.
 
   if(!(_.keys(badWordsFile.read()).length > 0)) {
-    badWordsFile.set(`${message.guild.id}.badWords`, []);
+    badWordsFile.set(`badWords`, []);
     badWordsFile.save();
   }
 
@@ -2365,17 +2370,13 @@ client.on('message', async message => {
             customCommands.set(args[1]+'.end', "");
             customCommands.save();
             message.channel.send(":white_check_mark: Command added successfully.");
-          } else {
-            syntaxErr(message, `commands_add`);
-          }
+          } else syntaxErr(message, `commands_add`);
         } if (args[0] === "remove") {
             if(args[1]) {
               customCommands.unset(args[1]);
               customCommands.save();
               message.channel.send(":white_check_mark: Command removed successfully.");
-            } else {
-              syntaxErr(message, `commands_remove`);
-            }
+            } else syntaxErr(message, `commands_remove`);
           }
         }
        if(args[0] === "list") {
@@ -2717,7 +2718,7 @@ client.on('message', async message => {
               }
             }
           }
-          var currentWords = badWordsFile.get(message.guild.id).badWords;
+          var currentWords = badWordsFile.get(`badWords`);
           if(currentWords.length > 0) {
             if (currentWords.some(word => newWords.includes(word))) {
               message.channel.send(`:x: One or more words are already on the list.`);
@@ -2725,7 +2726,7 @@ client.on('message', async message => {
             }
           }
           currentWords = currentWords.concat(newWords);
-          badWordsFile.set(`${message.guild.id}.badWords`, currentWords);
+          badWordsFile.set(`badWords`, currentWords);
           badWordsFile.save();
           message.channel.send(`:white_check_mark: Word(s) \`${newWords}\` added successfully.`);
         } else {
@@ -2751,7 +2752,7 @@ client.on('message', async message => {
               }
             }
           }
-          var currentWords = badWordsFile.get(message.guild.id).badWords;
+          var currentWords = badWordsFile.get(`badWords`);
           var numberOfElements = 0;
           for (var i = 0; i < newWords.length; i++) {
             for (var j = 0; j < currentWords.length; j++) {
@@ -2784,9 +2785,9 @@ client.on('message', async message => {
     }
     if(args[0] === "clear") {
       if(message.member.roles.some(role => ["Moderators"].includes(role.name))) {
-        if(badWordsFile.get(message.guild.id).badWords.length > 0) {
-          badWordsFile.unset(`${message.guild.id}`);
-          badWordsFile.set(`${message.guild.id}`.badWords, []);
+        if(badWordsFile.get(`badWords`).length > 0) {
+          badWordsFile.unset(`badWords`);
+          badWordsFile.set(`badWords`, []);
           badWordsFile.save();
           message.channel.send(`:white_check_mark: All bad words successfully removed.`);
         } else message.channel.send(`:x: No bad words could be found.`)
@@ -2794,7 +2795,7 @@ client.on('message', async message => {
     }
     if(args[0] === "list"){
       if(message.member.roles.some(role=>["Moderators"].includes(role.name))){
-        var currentWords = badWordsFile.get(message.guild.id).badWords.join(`\n`);
+        var currentWords = badWordsFile.get(`badWords`).join(`\n`);
         if(currentWords) {
           message.channel.send({embed: {
               color: config.color_info,
